Adds support for UMD modules.
Adds support to UMD (Universal Module Definition) for Bideo.js. The code is
based on the source code of underscodejs
(http://underscorejs.org/docs/underscore.html) because it was easier to implement
and required less changes in the source code rather than the definition of
UMDjs (https://github.com/umdjs/umd).

It supports CommonJS, AMD and global browser objects and includes simple tests
to verify it.